new new

+++ usart/my_math.c
@@ -1,6 +1,6 @@
 #include "my_math.h"
 
-int32_t sum_of_array(uint16_t* arr,uint8_t count){
+int32_t sum_of_array(uint8_t* arr,uint8_t count){
 	uint16_t sum = 0;
 	
 	for(uint8_t i = 0; i<count; i++)
